requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
arxiv==2.1.3
attrs==24.2.0
beautifulsoup4==4.12.3
black==24.8.0
Bottleneck==1.3.7
build==1.2.1
CacheControl==0.14.0
certifi==2024.7.4
cffi==1.17.0
cfgv==3.4.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
crashtest==0.4.1
cryptography==43.0.0
dataclasses-json==0.6.7
distlib==0.3.8
distro==1.9.0
dulwich==0.21.7
fastjsonschema==2.20.0
feedparser==6.0.11
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
greenlet==3.0.3
groq==0.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.6
identify==2.6.0
idna==3.7
installer==0.7.0
jaraco.classes==3.4.0
jeepney==0.8.0
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
keyring==24.3.1
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.34
langchain-text-splitters==0.2.2
langsmith==0.1.103
markdownify==0.13.1
MarkupSafe==2.1.5
marshmallow==3.22.0
more-itertools==10.4.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.3
nodeenv==1.9.1
numexpr==2.8.7
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
orjson==3.10.7
packaging==24.1
pandas==2.2.2
pathspec==0.12.1
pexpect==4.9.0
pgvector==0.3.2
pillow==10.4.0
pip==24.2
pip-autoremove==0.10.0
pkginfo==1.11.1
platformdirs==4.2.2
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
pre-commit==3.8.0
psycopg==3.2.1
psycopg-binary==3.2.1
psycopg-pool==3.2.2
ptyprocess==0.7.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pyproject_hooks==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
rapidfuzz==3.9.6
regex==2024.7.24
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.1
SecretStorage==3.3.3
sentence-transformers==3.0.1
setuptools==72.1.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.32
sympy==1.13.2
tenacity==8.5.0
threadpoolctl==3.5.0
tokenizers==0.19.1
tomlkit==0.13.2
torch==2.4.0
tqdm==4.66.4
transformers==4.44.2
triton==3.0.0
trove-classifiers==2024.7.2
types-requests==2.32.0.20240712
typing_extensions==4.12.2
typing-inspect==0.9.0
tzdata==2023.3
urllib3==2.2.2
virtualenv==20.26.3
wheel==0.43.0
yarl==1.9.4
